refactor web app configuration
