+++ web/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "mapwisefox-web"
-version = "0.2.0"
+version = "0.3.0"
 dependencies = [
     "arrow>=1.3.0",
     "authlib>=1.6.0",
