+++ web/src/mapwisefox/static/favicon.ico
