+++ web/src/mapwisefox/static/home.css
