+++ web/src/mapwisefox/static/style.css
