+++ web/src/mapwisefox/web/config/__init__.py
@@ -1,0 +1,4 @@
+from ._deps import settings
+from ._settings import AppSettings
+
+__all__ = ["settings", "AppSettings"]
