+++ web/src/mapwisefox/web/config/_deps.py
@@ -1,0 +1,8 @@
+from functools import lru_cache
+
+from ._settings import AppSettings
+
+
+@lru_cache()
+def settings() -> AppSettings:
+    return AppSettings()
