--- web/src/mapwisefox/web/static/__init__.py
@@ -1,3 +1,0 @@
-from pathlib import Path
-
-static_dir_path = Path(__file__).parent
